--- entities/小云米
@@ -1,2 +1,0 @@
-// here will be yummi
-{}
