+++ package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
     "@minecraft/server": "^1.4.0-beta.1.20.10-stable",
-    "@minecraft/server-gametest": "^1.0.0-beta.1.20.10-preview.21"
+    "@minecraft/server-gametest": "^1.0.0-beta.1.20.10-stable"
   }
 }
