+++ scripts/main/verifyDataBase.js
@@ -1,0 +1,16 @@
+import ScoreBase from "../lib/xboyPackage/scoreBase/rw";
+let ScoreBaseSnapshot = ScoreBase.GetObject();
+const checkScoreObjectExist = (ScoreObjectName) => !!Array.from(ScoreBaseSnapshot).find((ScoreObject) => { if (ScoreObjectName === ScoreObject.id)
+    return true; });
+const verif = function () {
+    ScoreBaseSnapshot = ScoreBase.GetObject();
+    ["##FlashPlayer##"].forEach((_) => {
+        checkScoreObjectExist(_)
+            ? console.error(_, "存在")
+            : (ScoreBase.NewObjectAsync('"' + _ + '"', '"' + _ + '"', "dummy"), console.error(_, "不存在但已创建"));
+    });
+    !!ScoreBase.AssPartic("##currentPID", ScoreBase.GetObject("##FlashPlayer##"))
+        ? console.error("数据存在==>", "##currentPID")
+        : (ScoreBase.AddPointsAsync('"' + '##currentPID' + '"', '"' + '##FlashPlayer##' + '"', "0"), console.error("数据不存在但已创建==>", "##xSkyLands##currentUID"));
+};
+export default verif;
