+++ 重写.md
@@ -1,0 +1,50 @@
+# 流程
+```
+    # 初始化 (function:init)
+
+    -检测0号ceyk(tag:init)实体以及坐标 
+     -有则跳过创建或修正坐标
+     -无则创建0号ceyk实体 实际坐标 (30000000 128 0) 并添加(tag:init)
+    -使用0号实体完成区域加载
+    -使用fill完成区域清理 (29999997 0 5 30000002 319 -1) 
+     * 待商榷改用getBlock
+    -执行gametest创建test环境 坐标 (30000000 128 0)
+```
+```
+    # 生成模拟玩家
+    
+    -未指定数字序号
+     -使用数字序号
+    -指定数字序号
+     -创建数字序号
+    -创建模拟玩家
+    -增加 (tag:init tag:xboySign)
+    -存入模拟玩家列表
+    -检测是否存在相同数字序号的ceyk实体
+     -有则将ceyk实体所记录物品与经验值同步至此模拟玩家->存在 (tag:init)
+      * 恢复物品与经验 
+      * 记分板，tag，药水效果计入待办事项
+     -无则生成ceyk实体并初始化
+    -记录模拟玩家创建位置 test相对位置(0 0 0)
+    -设置模拟玩家重生点至指定维度与坐标    
+    -传送模拟玩家至指定维度与坐标
+```
+```
+    # 生成常规ceyk并初始化
+
+    -使用指定数字序号
+    -若对应序号虚拟玩家存在 且不存在 (tag:init)
+     -同步虚拟玩家物品与经验值到此实体 每tick
+     * 此处逻辑实际相反 取所有模拟玩家列表 遍历并同步至对应ceyk实体
+     * 每tick应该没有性能压力吧
+```
+
+```
+    # 环境变量维护
+
+    -变量
+        xboySign = "#xboySimSign#"
+    -记分板
+        xboySign 储存模拟玩家与ceyk实体数字序号 0号ceyk的序号为当前最大序号
+
+```
